* Added support for changing the text of back button and customizing menu colors.

+++ SplitViewMenuUWP/SplitViewMenuUWP/Scenario2/Scenario2Main.xaml.cs
@@ -12,21 +12,21 @@ namespace SplitViewMenuUWP.Scenario2
             {
                 Label = "Page 1",
                 DestinationPage = typeof (Scenario2Page1),
-                Symbol = Symbol.Bookmarks,
+                Symbol = (int)Symbol.Bookmarks,
                 NotifyCount = 2
             });
             mainViewModel.MenuItems.Add(new Scenario2NavMenuIItem
             {
                 Label = "Page 2",
                 DestinationPage = typeof (Scenario2Page2),
-                Symbol = Symbol.Emoji,
+                Symbol = (int)Symbol.Emoji,
                 NotifyCount = 7
             });
             mainViewModel.MenuItems.Add(new Scenario2NavMenuIItem
             {
                 Label = "Page 3",
                 DestinationPage = typeof (Scenario2Page3),
-                Symbol = Symbol.RotateCamera,
+                Symbol = (int)Symbol.RotateCamera,
                 NotifyCount = 1
             });
             DataContext = mainViewModel;
